rm, status, clean impl. test fixes

+++ src/main/java/ru/spbau/Checkout.java
@@ -13,6 +13,7 @@ import java.util.Map;
 
 import static ru.spbau.Branch.checkIfBranchExist;
 import static ru.spbau.Commit.getCommit;
+import static ru.spbau.Init.hasInitialized;
 
 /**
  * Class implementing checkout command.
@@ -26,12 +27,13 @@ public class Checkout {
      *                                    the work with file system
      * @throws Vcs.VcsBranchNotFoundException when trying to access branch
      *                                    which doesn't exist.
-     * @throws Vcs.VcsIllegalStateException   when vcs can't perform command because of incorrect
+     * @throws Vcs.VcsIncorrectUsageException   when vcs can't perform command because of incorrect
      *                                    usage
      */
-    public static void checkoutBranch(String branchName) throws IOException, Vcs.VcsBranchNotFoundException, Vcs.VcsIllegalStateException {
+    public static void checkoutBranch(String branchName) throws IOException, Vcs.VcsBranchNotFoundException, Vcs.VcsIncorrectUsageException {
+        if (!hasInitialized()) throw new Vcs.VcsIncorrectUsageException(Vcs.getUninitializedRepoMessage());
         checkIfBranchExist(branchName);
-        String commitHash = FileSystem.getFirstLine(Vcs.getBranchesDir() + File.separator + branchName);
+        String commitHash = Branch.getBranchLastCommitHash(branchName);
         checkout(commitHash);
 
         FileSystem.writeStringToFile(Vcs.getHEAD(), branchName);
@@ -45,21 +47,25 @@ public class Checkout {
      *                                      the work with file system
      * @throws Vcs.VcsRevisionNotFoundException when trying to access revision
      *                                      which doesn't exist
-     * @throws Vcs.VcsIllegalStateException     when vcs can't perform command because of incorrect
+     * @throws Vcs.VcsIncorrectUsageException     when vcs can't perform command because of incorrect
      *                                      usage
      */
-    public static void checkoutRevision(String commitHash) throws IOException, Vcs.VcsRevisionNotFoundException, Vcs.VcsIllegalStateException {
+    public static void checkoutRevision(String commitHash) throws IOException, Vcs.VcsRevisionNotFoundException, Vcs.VcsIncorrectUsageException {
+        if (!hasInitialized()) throw new Vcs.VcsIncorrectUsageException(Vcs.getUninitializedRepoMessage());
         checkIfRevisionExist(commitHash);
         checkout(commitHash);
 
         FileSystem.writeStringToFile(Vcs.getBranchesDir() + File.separator + Branch.getHeadBranch(), commitHash);
     }
 
-    private static void checkout(String commitHash) throws IOException, Vcs.VcsIllegalStateException {
-        String lastCommitHash = FileSystem.getFirstLine(Vcs.getBranchesDir() + File.separator + Branch.getHeadBranch());
+    private static void checkout(String commitHash) throws IOException, Vcs.VcsIncorrectUsageException {
+        String lastCommitHash = Branch.getBranchLastCommitHash(Branch.getHeadBranch());
         if (lastCommitHash.equals(commitHash)) return;
         if (!FileSystem.getFirstLine(Vcs.getAddList()).equals("")) {
-            throw new Vcs.VcsIllegalStateException("You have several files were added, but haven't committed yet");
+            throw new Vcs.VcsIncorrectUsageException("You have several files were added, but haven't committed yet");
+        }
+        if (!FileSystem.getFirstLine(Vcs.getRmList()).equals("")) {
+            throw new Vcs.VcsIncorrectUsageException("You have several files were removed, but haven't committed yet");
         }
         deleteCommittedFiles(lastCommitHash);
         Collection<String> restored = new HashSet<>();
@@ -68,7 +74,7 @@ public class Checkout {
 
     private static void deleteCommittedFiles(String commitHash) throws IOException {
         VcsCommit commit = getCommit(commitHash);
-        for (Map.Entry<String, String> entry : commit.getChildren().entrySet()) {
+        for (Map.Entry<String, String> entry : commit.getChildrenAdd().entrySet()) {
             FileSystem.deleteIfExists(entry.getKey());
         }
         if (!commit.getPrevCommitHash().equals(Vcs.getInitialCommitPrevHash())) {
@@ -78,12 +84,15 @@ public class Checkout {
 
     private static void restore(String commitHash, Collection<String> restored) throws IOException {
         VcsCommit commit = getCommit(commitHash);
-        for (Map.Entry<String, String> entry : commit.getChildren().entrySet()) {
+        for (Map.Entry<String, String> entry : commit.getChildrenAdd().entrySet()) {
             if (!restored.contains(entry.getKey())) {
                 VcsBlob blob = (VcsBlob) VcsObject.readFromJson(Vcs.getObjectsDir() + File.separator + entry.getValue(), VcsBlob.class);
                 FileSystem.writeBytesToFile(entry.getKey(), blob.getContent());
                 restored.add(entry.getKey());
             }
+        }
+        for (String fileRm : commit.getChildrenRm()) {
+            restored.add(fileRm);
         }
         if (!commit.getPrevCommitHash().equals(Vcs.getInitialCommitPrevHash())) {
             restore(commit.getPrevCommitHash(), restored);
